// src/app/models/flights.modals.ts
export class Flights {
    'id': number;
    'flightId': string;
    'imgUrl': string;
    'from': string;
    'to': string;
    'departureTime': string;
    'arrivalTime': string;
    'duration': string;
    'ancillaryServices': Array<{
    }>;
    'shoppingItems': Array<{
    }>;
    'specialMeals': Array<{
      }>;
  }